requirements.txt
beautifulsoup4
ebooklib
lxml
Pillow
PyQt6
pywebview
requests
tqdm 